Seperating the player class
